split manga_translator.py into multiple files
