+++ manga_translator/mode/local.py
@@ -1,0 +1,201 @@
+import os
+from typing import Union, List
+
+from PIL import Image
+
+from manga_translator import MangaTranslator, logger, Context, TranslationInterrupt
+from ..save import save_result
+from ..translators import (
+    LanguageUnsupportedException,
+    dispatch as dispatch_translation,
+)
+from ..utils import natural_sort, replace_prefix, get_color_name, rgb2hex
+
+
+class MangaTranslatorLocal(MangaTranslator):
+    async def translate_path(self, path: str, dest: str = None, params: dict[str, Union[int, str]] = None):
+        """
+        Translates an image or folder (recursively) specified through the path.
+        """
+        if not os.path.exists(path):
+            raise FileNotFoundError(path)
+        path = os.path.abspath(os.path.expanduser(path))
+        dest = os.path.abspath(os.path.expanduser(dest)) if dest else ''
+        params = params or {}
+
+        # Handle format
+        file_ext = params.get('format')
+        if params.get('save_quality', 100) < 100:
+            if not params.get('format'):
+                file_ext = 'jpg'
+            elif params.get('format') != 'jpg':
+                raise ValueError('--save-quality of lower than 100 is only supported for .jpg files')
+
+        if os.path.isfile(path):
+            # Determine destination file path
+            if not dest:
+                # Use the same folder as the source
+                p, ext = os.path.splitext(path)
+                _dest = f'{p}-translated.{file_ext or ext[1:]}'
+            elif not os.path.basename(dest):
+                p, ext = os.path.splitext(os.path.basename(path))
+                # If the folders differ use the original filename from the source
+                if os.path.dirname(path) != dest:
+                    _dest = os.path.join(dest, f'{p}.{file_ext or ext[1:]}')
+                else:
+                    _dest = os.path.join(dest, f'{p}-translated.{file_ext or ext[1:]}')
+            else:
+                p, ext = os.path.splitext(dest)
+                _dest = f'{p}.{file_ext or ext[1:]}'
+            await self.translate_file(path, _dest, params)
+
+        elif os.path.isdir(path):
+            # Determine destination folder path
+            if path[-1] == '\\' or path[-1] == '/':
+                path = path[:-1]
+            _dest = dest or path + '-translated'
+            if os.path.exists(_dest) and not os.path.isdir(_dest):
+                raise FileExistsError(_dest)
+
+            translated_count = 0
+            for root, subdirs, files in os.walk(path):
+                files = natural_sort(files)
+                dest_root = replace_prefix(root, path, _dest)
+                os.makedirs(dest_root, exist_ok=True)
+                for f in files:
+                    if f.lower() == '.thumb':
+                        continue
+
+                    file_path = os.path.join(root, f)
+                    output_dest = replace_prefix(file_path, path, _dest)
+                    p, ext = os.path.splitext(output_dest)
+                    output_dest = f'{p}.{file_ext or ext[1:]}'
+
+                    if await self.translate_file(file_path, output_dest, params):
+                        translated_count += 1
+            if translated_count == 0:
+                logger.info('No further untranslated files found. Use --overwrite to write over existing translations.')
+            else:
+                logger.info(f'Done. Translated {translated_count} image{"" if translated_count == 1 else "s"}')
+
+    async def translate_file(self, path: str, dest: str, params: dict):
+        if not params.get('overwrite') and os.path.exists(dest):
+            logger.info(
+                f'Skipping as already translated: "{dest}". Use --overwrite to overwrite existing translations.')
+            await self._report_progress('saved', True)
+            return True
+
+        logger.info(f'Translating: "{path}"')
+
+        # Turn dict to context to make values also accessible through params.<property>
+        params = params or {}
+        ctx = Context(**params)
+        self._preprocess_params(ctx)
+
+        attempts = 0
+        while ctx.attempts == -1 or attempts < ctx.attempts + 1:
+            if attempts > 0:
+                logger.info(f'Retrying translation! Attempt {attempts}'
+                            + (f' of {ctx.attempts}' if ctx.attempts != -1 else ''))
+            try:
+                return await self._translate_file(path, dest, ctx)
+
+            except TranslationInterrupt:
+                break
+            except Exception as e:
+                if isinstance(e, LanguageUnsupportedException):
+                    await self._report_progress('error-lang', True)
+                else:
+                    await self._report_progress('error', True)
+                if not self.ignore_errors and not (ctx.attempts == -1 or attempts < ctx.attempts):
+                    raise
+                else:
+                    logger.error(f'{e.__class__.__name__}: {e}',
+                                 exc_info=e if self.verbose else None)
+            attempts += 1
+        return False
+
+    async def _translate_file(self, path: str, dest: str, ctx: Context) -> bool:
+        if path.endswith('.txt'):
+            with open(path, 'r') as f:
+                queries = f.read().split('\n')
+            translated_sentences = \
+                await dispatch_translation(ctx.translator, queries, ctx.use_mtpe, ctx,
+                                           'cpu' if self._gpu_limited_memory else self.device)
+            p, ext = os.path.splitext(dest)
+            if ext != '.txt':
+                dest = p + '.txt'
+            logger.info(f'Saving "{dest}"')
+            with open(dest, 'w') as f:
+                f.write('\n'.join(translated_sentences))
+            return True
+
+        # TODO: Add .gif handler
+
+        else:  # Treat as image
+            try:
+                img = Image.open(path)
+                img.verify()
+                img = Image.open(path)
+            except Exception:
+                logger.warn(f'Failed to open image: {path}')
+                return False
+
+            ctx = await self.translate(img, ctx)
+            result = ctx.result
+
+            # Save result
+            if ctx.skip_no_text and not ctx.text_regions:
+                logger.debug('Not saving due to --skip-no-text')
+                return True
+            if result:
+                logger.info(f'Saving "{dest}"')
+                save_result(result, dest, ctx)
+                await self._report_progress('saved', True)
+
+                if ctx.save_text or ctx.save_text_file or ctx.prep_manual:
+                    if ctx.prep_manual:
+                        # Save original image next to translated
+                        p, ext = os.path.splitext(dest)
+                        img_filename = p + '-orig' + ext
+                        img_path = os.path.join(os.path.dirname(dest), img_filename)
+                        img.save(img_path, quality=ctx.save_quality)
+                    if ctx.text_regions:
+                        self._save_text_to_file(path, ctx)
+                return True
+        return False
+
+    def _save_text_to_file(self, image_path: str, ctx: Context):
+        cached_colors = []
+
+        def identify_colors(fg_rgb: List[int]):
+            idx = 0
+            for rgb, _ in cached_colors:
+                # If similar color already saved
+                if abs(rgb[0] - fg_rgb[0]) + abs(rgb[1] - fg_rgb[1]) + abs(rgb[2] - fg_rgb[2]) < 50:
+                    break
+                else:
+                    idx += 1
+            else:
+                cached_colors.append((fg_rgb, get_color_name(fg_rgb)))
+            return idx + 1, cached_colors[idx][1]
+
+        s = f'\n[{image_path}]\n'
+        for i, region in enumerate(ctx.text_regions):
+            fore, back = region.get_font_colors()
+            color_id, color_name = identify_colors(fore)
+
+            s += f'\n-- {i + 1} --\n'
+            s += f'color: #{color_id}: {color_name} (fg, bg: {rgb2hex(*fore)} {rgb2hex(*back)})\n'
+            s += f'text:  {region.text}\n'
+            s += f'trans: {region.translation}\n'
+            for line in region.lines:
+                s += f'coords: {list(line.ravel())}\n'
+        s += '\n'
+
+        text_output_file = ctx.text_output_file
+        if not text_output_file:
+            text_output_file = os.path.splitext(image_path)[0] + '_translations.txt'
+
+        with open(text_output_file, 'a', encoding='utf-8') as f:
+            f.write(s)
